feat: implement validation for purchase request and refactor input handling
