+++ backend/internal/app/handler/purchases.go
@@ -1,7 +1,6 @@
 package handler
 
 import (
-	"fmt"
 	"net/http"
 	"strconv"
 
@@ -11,76 +10,7 @@ import (
 	response "github.com/potibm/kasseapparat/internal/app/response"
 	"github.com/potibm/kasseapparat/internal/app/service"
 	"github.com/potibm/kasseapparat/internal/app/utils"
-	"github.com/shopspring/decimal"
 )
-
-type PurchaseListItemRequest struct {
-	ID             int  `binding:"required" form:"ID"`
-	AttendedGuests uint `binding:"required" form:"attendedGuests"`
-}
-
-type PurchaseCartRequest struct {
-	ID        int                       `binding:"required"      form:"ID"`
-	Quantity  int                       `binding:"required"      form:"quantity"`
-	ListItems []PurchaseListItemRequest `binding:"required,dive" form:"listItems"`
-}
-
-type PurchaseRequest struct {
-	TotalNetPrice   decimal.Decimal       `binding:"required"      form:"totalNetPrice"`
-	TotalGrossPrice decimal.Decimal       `binding:"required"      form:"totalGrossPrice"`
-	Cart            []PurchaseCartRequest `binding:"required,dive" form:"cart"`
-	PaymentMethod   string                `binding:"required"      form:"paymentMethod"`
-}
-
-func (req PurchaseRequest) ToInput() (service.PurchaseInput, error) {
-	if req.TotalNetPrice.IsNegative() || req.TotalGrossPrice.IsNegative() {
-		return service.PurchaseInput{}, fmt.Errorf("total price must not be negative")
-	}
-
-	if len(req.Cart) == 0 {
-		return service.PurchaseInput{}, fmt.Errorf("cart must not be empty")
-	}
-
-	seen := make(map[int]struct{})
-
-	var input service.PurchaseInput
-
-	input.PaymentMethod = req.PaymentMethod
-	input.TotalNetPrice = req.TotalNetPrice
-	input.TotalGrossPrice = req.TotalGrossPrice
-
-	for _, cart := range req.Cart {
-		if cart.Quantity < 1 {
-			return service.PurchaseInput{}, fmt.Errorf("quantity must be at least 1")
-		}
-
-		if _, ok := seen[cart.ID]; ok {
-			return service.PurchaseInput{}, fmt.Errorf("duplicate product ID: %d", cart.ID)
-		}
-
-		seen[cart.ID] = struct{}{}
-
-		item := service.PurchaseCartItem{
-			ID:       cart.ID,
-			Quantity: cart.Quantity,
-		}
-
-		for _, li := range cart.ListItems {
-			if li.AttendedGuests < 1 {
-				return service.PurchaseInput{}, fmt.Errorf("attendedGuests must be at least 1")
-			}
-
-			item.ListItems = append(item.ListItems, service.ListItemInput{
-				ID:             li.ID,
-				AttendedGuests: int(li.AttendedGuests),
-			})
-		}
-
-		input.Cart = append(input.Cart, item)
-	}
-
-	return input, nil
-}
 
 func (handler *Handler) DeletePurchase(c *gin.Context) {
 	executingUserObj, err := handler.getUserFromContext(c)
@@ -120,11 +50,12 @@ func (handler *Handler) PostPurchases(c *gin.Context) {
 		return
 	}
 
-	input, err := req.ToInput()
-	if err != nil {
+	if err := req.Validate(); err != nil {
 		_ = c.Error(ExtendHttpErrorWithDetails(InvalidRequest, err.Error()))
 		return
 	}
+
+	input := req.ToInput()
 
 	purchaseService := service.NewPurchaseService(handler.repo, &handler.mailer, handler.decimalPlaces)
